+++ internal/builder/configmap.go
@@ -12,7 +12,7 @@ import (
 
 	yamlv3 "gopkg.in/yaml.v3"
 
-	pgpilotv1 "github.com/jamalshahverdiyev/pgpilot-operator/api/v1"
+	pgpilotv1 "github.com/jamalshahverdiev/pgpilot-operator/api/v1"
 )
 
 const (