+++ internal/builder/service_test.go
@@ -3,7 +3,7 @@ package builder
 import (
 	"testing"
 
-	pgpilotv1 "github.com/jamalshahverdiyev/pgpilot-operator/api/v1"
+	pgpilotv1 "github.com/jamalshahverdiev/pgpilot-operator/api/v1"
 )
 
 func TestBuildService_BasicShape(t *testing.T) {