+++ test/e2e/e2e_suite_test.go
@@ -28,7 +28,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	"github.com/jamalshahverdiyev/pgpilot-operator/test/utils"
+	"github.com/jamalshahverdiev/pgpilot-operator/test/utils"
 )
 
 var (